Extract active-context sanitizers into sanitize.py (WS5 seam 2)
Move the pure active-context sanitization helpers out of LCMEngine into a new
`sanitize.py` module:
- `_contains_sensitive_redaction`, `_structured_part_text`,
  `_structured_part_has_visible_assistant_content`,
  `_assistant_message_has_visible_content`, `_strip_structured_text_part`,
  `_sanitize_active_assistant_content`, `_clean_active_assistant_message`,
  `_should_drop_active_assistant_message`, and the
  `_VISIBLE_TEXT_PART_TYPES` / `_INTERNAL_ASSISTANT_PART_TYPES` constants.

These are pure functions that shape the active replay context (strip
internal/reasoning assistant content, decide visible-content, detect sensitive
redaction markers); they never touch stored rows. engine.py imports the three
with external callers and calls them as module functions; the remaining
call sites were the cluster's own internal cross-calls, which move with it.

Behaviour-preserving: bodies are moved verbatim -- the only edits are dropping
the `@staticmethod`/`@classmethod` decorators and `cls`/`self` params and
rewriting the intra-cluster `cls._x(...)` calls to bare module calls. No import
cycle: the module depends only on `escalation._strip_reasoning_blocks` and
itself. The stateful/config-coupled sanitizers (`_sanitize_active_context_messages`
orchestrator, `_sanitize_tool_pairs`, the preserved-objective helpers, and
`_redact_active_replay_messages`) intentionally stay on the engine.
